Fix the degree transformation problem in edge detection
When the program tried to get the gradient angle with Sobel X and
Y, the function it used, numpy.arctan2(), returns values between
-pi/2 and pi/2.
However, when the program make masks with degrees, it uses only
positive values between 0 and 360 as usual.
The original method to keep the program working is add 180 to the
returned values, but actually, the negative values should be
increased by 360 to be on the original place but with right and
positive values while, on the other hand, the positive should just
stay on the original place with the original values.
Therefore, the new method is just add all degress with 360 and
modulo 360. In that case, all values can obtain their new and right
degrees.

Also, when the program wanted to get the positive values from Sobel
X and Y, it turned every value to their square root, but now it
just put them into abs() for the efficiency.

Finally, the learning rate in VGG19.py got increased to 0.01 for
better performance on validation accuracy.